move static route definitions to features
