+++ app/features/counter/routes.go
@@ -1,6 +1,10 @@
 package counter
 
 import (
+	"net/http"
+	"northstar/app/features/counter/web"
+
+	"github.com/benbjohnson/hashfs"
 	"github.com/go-chi/chi/v5"
 	"github.com/gorilla/sessions"
 )
@@ -8,6 +12,7 @@ import (
 func SetupRoutes(router chi.Router, sessionStore sessions.Store) error {
 	handlers := NewHandlers(sessionStore)
 
+	router.Handle("/counter/static/*", http.StripPrefix("/counter", hashfs.FileServer(web.StaticSys)))
 	router.Get("/counter", handlers.CounterPage)
 	router.Get("/counter/data", handlers.CounterData)
