+++ app/features/index/routes.go
@@ -1,8 +1,11 @@
 package index
 
 import (
+	"net/http"
 	"northstar/app/features/index/services"
+	"northstar/app/features/index/web"
 
+	"github.com/benbjohnson/hashfs"
 	"github.com/delaneyj/toolbelt/embeddednats"
 	"github.com/go-chi/chi/v5"
 	"github.com/gorilla/sessions"
@@ -16,6 +19,7 @@ func SetupRoutes(router chi.Router, store sessions.Store, ns *embeddednats.Serve
 
 	handlers := NewHandlers(todoService)
 
+	router.Handle("/index/static/*", http.StripPrefix("/index", hashfs.FileServer(web.StaticSys)))
 	router.Get("/", handlers.IndexPage)
 
 	router.Route("/api", func(apiRouter chi.Router) {
